+++ tfts/models/auto_config.py
@@ -3,7 +3,6 @@
 # @author: Longxing Tan, tanlongxing888@163.com
 """AutoConfig to set up models custom config"""
 
-import collections
 
 from tfts.models.autoformer import params as autoformer_params
 from tfts.models.bert import params as bert_params
@@ -56,28 +55,3 @@ class AutoConfig:
 
     def save_config(self):
         return
-
-
-class BaseConfig(object):
-    def to_dict(self):
-        output_dict = {}
-        for key, value in self.__dict__.items():
-            output_dict[key] = value
-        return flatten_dict(output_dict)
-
-
-def flatten_dict(nested, sep="/"):
-    """Flatten dictionary and concatenate nested keys with separator."""
-
-    def rec(nest, prefix, into):
-        for k, v in nest.items():
-            if sep in k:
-                raise ValueError(f"separator '{sep}' not allowed to be in key '{k}'")
-            if isinstance(v, collections.Mapping):
-                rec(v, prefix + k + sep, into)
-            else:
-                into[prefix + k] = v
-
-    flat = {}
-    rec(nested, "", flat)
-    return flat
