+++ tfts/models/base.py
@@ -1,0 +1,32 @@
+from abc import ABC, abstractmethod
+import collections
+
+
+class BaseModel(ABC):
+    def __init__(self):
+        pass
+
+
+class BaseConfig(ABC):
+    def to_dict(self):
+        output_dict = {}
+        for key, value in self.__dict__.items():
+            output_dict[key] = value
+        return flatten_dict(output_dict)
+
+
+def flatten_dict(nested, sep="/"):
+    """Flatten dictionary and concatenate nested keys with separator."""
+
+    def rec(nest, prefix, into):
+        for k, v in nest.items():
+            if sep in k:
+                raise ValueError(f"separator '{sep}' not allowed to be in key '{k}'")
+            if isinstance(v, collections.Mapping):
+                rec(v, prefix + k + sep, into)
+            else:
+                into[prefix + k] = v
+
+    flat = {}
+    rec(nested, "", flat)
+    return flat
