+++ tfts/models/rwkv.py
@@ -1,0 +1,4 @@
+"""
+`RWKV: Reinventing RNNs for the Transformer Era
+<https://arxiv.org/abs/2305.13048>`_
+"""
